ASP.NET CORE WEB API
In the context of this project, ASP.NET Core Web was utilized to establish the foundation for 
the web application. Both The backend logic and User Interface was implemented using 
ASP.NET Core Web. This framework facilitated the creation of robust and efficient APIs 
through which data could be seamlessly exchanged between the frontend and backend 
components of the application.