ASP.NET CORE WEB API
The structure of 
my ASP.NET web project is 
organized into distinct folders, 
each serving a specific purpose 
to ensure a systematic and 
modularized development 
approach. The "Controllers" 
folder houses the controllers of 
the Model-View-Controller 
(MVC) architectural pattern, 
responsible for handling the 
application's logic and 
managing the flow of data 
between the model and the 
views. In the "Models" folder, 
the essential data structures and 
business logic of the 
application are encapsulated, 
promoting a separation of 
concerns. The "Migrations" 
folder is integral to the database 
schema management, housing 
scripts that facilitate the 
evolution of the database as the 
application evolves. Views, 
representing the user interface, 
are stored in the "Views" 
folder, allowing for a clear 
distinction between the 
backend and frontend 
components. The 
"appsettings.json" file contains 
configuration settings for the application, providing a centralized location for managing various 
parameters. Finally, the "Program.cs" file serves as the entry point of the application, 
orchestrating the startup sequence and configuration. This folder structure adheres to best 
practices in software development, ensuring clarity, maintainability, and scalability throughout 
the development lifecycle